Improve analyze for zedstore.
Physical number of blocks doesn't provide way to correctly sample
blocks for zedstore. Hence, adding new table API to get logical number
of blocks. Use this interface for analyze and estimate relsize
functions.

For heap, get logical and physical number of blocks uses same function
hence returns same value.

For zedstore, logical number of blocks is calculated based on largest
TID allocated and convert it to block number. This helps to get
correct number of reltuples at least in pg_class.

With this change now join_hash test failuring earlier for zedstore
passes as uses correct number of buckets and don't spill unnecessarily
for zedstore.

Not sure this is right way to go but showcases something which might
be needed.